requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Data augmentation
albumentations>=1.3.0
opencv-python>=4.7.0

# Configuration
PyYAML>=6.0

# Visualization and metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0

# Optional: Experiment tracking
wandb>=0.15.0

# Optional: Image loading from URLs
requests>=2.31.0

# Development
pytest>=7.3.0

